Upgrading phpstan level to 2

+++ Kata/Y2026/Q2/CarParkEscape.php
@@ -67,7 +67,7 @@ function escape(array $carpark): array
                 }
             } else {
                 if ($arrayCount != $index + 1) {
-                    $return[array_key_last($return)] = "D" . (str_split($return[array_key_last($return)])[1] + 1);
+                    $return[array_key_last($return)] = "D" . ((int)str_split($return[array_key_last($return)])[1] + 1);
                 }
             }
